+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Line wrap the file at 100 chars.                                              Th
 #### Linux
 - Make offline monitor aware of routing table changes.
 - Assign local DNS servers to more appropriate interfaces when using systemd-resolved.
+- Disable DNS over TLS for tunnel's DNS config when using systemd-resolved.
 
 #### Windows
 - Fix failure to restart the daemon when resuming from "fast startup" hibernation.